% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biGPD.R
\name{BiGPDApproachReturnLevels}
\alias{BiGPDApproachReturnLevels}
\title{Example of biGPD approach to get return levels from return period. Compute the two return levels with equal univariate probability from a given bivariate return period..}
\usage{
BiGPDApproachReturnLevels(
  data,
  returnPeriod,
  EGPDtypes,
  initParams1,
  initParams2,
  probaQuantile,
  nbDaysPerYear,
  nbYears,
  h,
  Dparam,
  probaOccurrence
)
}
\arguments{
\item{data}{A dataframe with two columns, one for each time series.}

\item{returnPeriod}{An integer, in years.}

\item{EGPDtypes}{A vector of size 2, one for each time series. Values are 1 or 4. See Naveau et al. (2016) and the package mev for further information.}

\item{initParams1}{A vector of parameters to initialize the estimation for the first time series. For type 1, it is (Kappa, Sigma, Xi), for type 4, it is (Prob, Kappa, Delta, Sigma, Xi).}

\item{initParams2}{A vector of parameters to initialize the estimation for the second time series. For type 1, it is (Kappa, Sigma, Xi), for type 4, it is (Prob, Kappa, Delta, Sigma, Xi).}

\item{probaQuantile}{A float between 0 and 1. Data above the thresholds of this probability are used to compute the bivariate exceedence probability. The value has a small impact over the probability. It should be close to 1, 0.95 is a classical value. The same value is used for both margins for simplicity.}

\item{nbDaysPerYear}{The number of days considered per year (integer).}

\item{nbYears}{The number of distinct years. Default value is 1.}

\item{h}{The parameter of non-concurrence (integer).}

\item{Dparam}{Cf documentation of the dgaps function of the exdex package. Default value is 3.}

\item{probaOccurrence}{The probability that the values are reached before the return period time. Default value to 0.63.}
}
\value{
A list with, in that order: the return level of first variable, the return level of the second variable, chi and chiBarre.
}
\description{
Example of biGPD approach to get return levels from return period. Compute the two return levels with equal univariate probability from a given bivariate return period..
}
