% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{BivariateDeclustering}
\alias{BivariateDeclustering}
\title{Decluster bivariate data.}
\usage{
BivariateDeclustering(
  data1,
  data2,
  nbDaysPerYear,
  nbYears,
  thresholds,
  blockSize,
  logic
)
}
\arguments{
\item{data1}{A vector of data for the first variable.}

\item{data2}{A vector of data for the second variable.}

\item{nbDaysPerYear}{The number of days considered per year (integer).}

\item{nbYears}{The number of distinct years.}

\item{thresholds}{A vector of size 2. The first threshold is for the first variable, and the second value is for the second variable. The threshold is the value above which data are supposed to follow a GPD.}

\item{blockSize}{The size of the blocks used for declustering.}

\item{logic}{Either "AND" or "OR". Couples of bloxk maxima are retained if they both exceed their respective threshold in case of "AND", and if at least one of them exceed its respective threshold in case of "OR". Default value to "AND".}
}
\value{
A dataframe with two columns, one for each variable, of declustered data following the logic parameter
}
\description{
Decluster bivariate data.
}
