% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biGPD.R
\name{BivariateExceedenceProbability2}
\alias{BivariateExceedenceProbability2}
\title{Compute the bivariate excess probability Fbar(x1,x2).}
\usage{
BivariateExceedenceProbability2(probas, empCDF, probaQuantile, FU1U2)
}
\arguments{
\item{probas}{A vector of size 2, with the 2 univariate probabilities. The two probabilities are likely close to 1.}

\item{empCDF}{The empirical cdf of Delta.}

\item{probaQuantile}{Data above the thresholds of this probability are used to compute the probability. The value has a small impact over the probability. It should be close to 1, 0.95 is a classical value. The same value is used for both margins for simplicity.}

\item{FU1U2}{The probability to be below both thresholds (quantiles of probability probaQuantile).}
}
\value{
The bivariate excess probability.
}
\description{
Compute the bivariate excess probability Fbar(x1,x2).
}
