% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biGPD.R
\name{BivariateExtremalIndex}
\alias{BivariateExtremalIndex}
\title{Bivariate extremal index}
\usage{
BivariateExtremalIndex(data1, data2, probaQuantile, probas, nbYears, Dparam)
}
\arguments{
\item{data1}{A vector of data for which the extremal index will be estimated.}

\item{data2}{A vector of data for which their extremal index will be estimated.}

\item{probaQuantile}{Data above the threshold of this probability are used to estimate the extremal index. The value has an impact over the estimation. It should be close to 1, 0.95 is a classical value.}

\item{probas}{A vector of probability at which the bivariate extermal index is estimated.}

\item{nbYears}{The number of distinct years. Default value is 1.}

\item{Dparam}{Cf documentation of the dgaps function of the exdex package. Default value is 3.}
}
\value{
The univariate extremal index.
}
\description{
Bivariate extremal index
}
