% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{BivariateReturnPeriodCopula}
\alias{BivariateReturnPeriodCopula}
\title{Calculate the bivariate return period with the copula approach.}
\usage{
BivariateReturnPeriodCopula(
  GPDprobas,
  h,
  extremalIndexes,
  probaQuantile,
  copula,
  FU1U2,
  nbDaysPerYear,
  nbYears,
  Dparam,
  probaOccurrence
)
}
\arguments{
\item{h}{The parameter of non-concurrence (integer).}

\item{extremalIndexes}{A list of size 3, with the extremal index for the first variable in first, the one for the second variable in second and the bivaraite extremal index in third.}

\item{probaQuantile}{The probabilityof the value above which data are supposed to follow a GPD. A common value is 0.95.}

\item{copula}{The copula family. The short name from the VineCopula package is expected.}

\item{FU1U2}{The probability to be below both thresholds (quantiles of probability probaQuantile).}

\item{nbDaysPerYear}{The number of days considered per year (integer).}

\item{nbYears}{The number of distinct years. Default value is 1.}

\item{Dparam}{Cf documentation of the dgaps function of the exdex package. Default value is 3.}

\item{probaOccurrence}{The probability that the values are reached before the return period time. Default value to 0.63.}

\item{returnLevels}{A vector of size 2, one for each time series. Return periods correspond to the return levels.}

\item{GPDParam1}{The parameters of the GPD for the first variable, in a list with the following order: c(threshold, sigma, xi).}

\item{GPDParam2}{The parameters of the GPD for the second variable, in a list with the following order: c(threshold, sigma, xi).}
}
\value{
The bivariate return period with the copula approach.
}
\description{
Calculate the bivariate return period with the copula approach.
}
