% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{CopulaApproach}
\alias{CopulaApproach}
\title{Example of copula approach. Compute univariate and bivariate return periods from 2 time series.}
\usage{
CopulaApproach(
  data,
  returnLevels,
  probaQuantile,
  nbDaysPerYear,
  nbYears,
  h,
  blockSizes,
  Dparam,
  probaOccurrence,
  logic
)
}
\arguments{
\item{data}{A dataframe with two columns, one for each time series.}

\item{returnLevels}{A vector of size 2, one for each time series. Return periods correspond to the return levels.}

\item{probaQuantile}{The probabilityof the value above which data are supposed to follow a GPD. A common value is 0.95.}

\item{nbDaysPerYear}{The number of days considered per year (integer).}

\item{nbYears}{The number of distinct years. Default value is 1.}

\item{h}{The parameter of non-concurrence (integer).}

\item{blockSizes}{Vector of integers of size 3. The sizes of the blocks considered for the declustering, with the univariates first and the bivariate block size at the end.}

\item{Dparam}{Cf documentation of the dgaps function of the exdex package. Default value is 3.}

\item{probaOccurrence}{The probability that the values are reached before the return period time. Default value to 0.63.}

\item{logic}{Either "AND" or "OR". Couples of bloxk maxima are retained if they both exceed their respective threshold in case of "AND", and if at least one of them exceed its respective threshold in case of "OR". Default value to "AND".}
}
\value{
A list with, in that order: the return period of first variable, the return period of the second variable, the bivariate return period, the non-concurrent joint excess probability, chi and chiBarre.
}
\description{
Example of copula approach. Compute univariate and bivariate return periods from 2 time series.
}
