% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biGPD.R
\name{EGPDtoExp}
\alias{EGPDtoExp}
\title{Transform data assumed to follow an EGPD distribution to a unit exponential distribution.}
\usage{
EGPDtoExp(data, EGPDtype, EGPDParams, probaZero)
}
\arguments{
\item{data}{A vector of data with the zeros filtered.}

\item{EGPDtype}{1 or 4. See Naveau et al. (2016) and the package mev for further information.}

\item{EGPDParams}{The EGPD parameters. For type 1, it is (Kappa, Sigma, Xi), for type 4, it is (Prob, Kappa, Delta, Sigma, Xi).}

\item{probaZero}{The probability to be exactly zero.}
}
\value{
A data vector.
}
\description{
Transform data assumed to follow an EGPD distribution to a unit exponential distribution.
}
