% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biGPD.R
\name{EstimateEGPDParameters}
\alias{EstimateEGPDParameters}
\title{Estimate the parameters of the EGPD with maximul likelihood (mle).}
\usage{
EstimateEGPDParameters(data, EGPDtype, initParam)
}
\arguments{
\item{data}{A vector of data with the zeros filtered.}

\item{EGPDtype}{1 or 4. See Naveau et al. (2016) and the package mev for further information.}

\item{initParam}{A vector of parameters to initialize the estimation. For type 1, it is (Kappa, Sigma, Xi), for type 4, it is (Prob, Kappa, Delta, Sigma, Xi).}
}
\value{
The parameters of the EGPD in the following order: for type 1, (Kappa, Sigma, Xi) and for type 4, (Prob, Kappa, Delta, Sigma, Xi).
}
\description{
Estimate the parameters of the EGPD with maximul likelihood (mle).
}
