% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{EstimateGPDParameters}
\alias{EstimateGPDParameters}
\title{Estimate the parameters of the GPD}
\usage{
EstimateGPDParameters(data, threshold)
}
\arguments{
\item{data}{A vector of data.}

\item{threshold}{The value above which data are supposed to follow a GPD.}
}
\value{
The estimated parameters of the GPD, in a list with the following order: c(threshold, sigma, xi).
}
\description{
Estimate the parameters of the GPD
}
