% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{GetGPDproba}
\alias{GetGPDproba}
\title{Compute the probability to be under the return level value with a GPD.}
\usage{
GetGPDproba(returnLevel, GPDParam)
}
\arguments{
\item{returnLevel}{The value at which the cdf is computed.}

\item{GPDParam}{The parameters of the GPD, in a list with the following order: c(threshold, sigma, xi).}
}
\value{
The cdf value at the return level value for a GPD.
}
\description{
Compute the probability to be under the return level value with a GPD.
}
