% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{GetGPDvalue}
\alias{GetGPDvalue}
\title{Compute the value corresponding to the cdf probability with a GPD.}
\usage{
GetGPDvalue(GPDproba, GPDParam)
}
\arguments{
\item{GPDproba}{The probability at which the inverse cdf is computed.}

\item{GPDParam}{The parameters of the GPD, in a list with the following order: c(threshold, sigma, xi).}
}
\value{
The value at which the cdf equals proba for a GPD.
}
\description{
Compute the value corresponding to the cdf probability with a GPD.
}
