% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnLevels.R
\name{GetReturnLevels}
\alias{GetReturnLevels}
\title{Select either the biGPD approach or the copula approach to get return levels from a given bivariate return period.}
\usage{
GetReturnLevels(
  data,
  returnPeriod,
  EGPDtypes,
  initParams1,
  initParams2,
  probaQuantile,
  nbDaysPerYear,
  nbYears,
  h,
  blockSizes,
  Dparam,
  probaOccurrence,
  logic
)
}
\arguments{
\item{data}{A dataframe with two columns, one for each time series.}

\item{returnPeriod}{An integer, in years.}

\item{EGPDtypes}{A vector of size 2, one for each time series. Values are 1 or 4. See Naveau et al. (2016) and the package mev for further information.}

\item{initParams1}{A vector of parameters to initialize the estimation for the first time series. For type 1, it is (Kappa, Sigma, Xi), for type 4, it is (Prob, Kappa, Delta, Sigma, Xi).}

\item{initParams2}{A vector of parameters to initialize the estimation for the second time series. For type 1, it is (Kappa, Sigma, Xi), for type 4, it is (Prob, Kappa, Delta, Sigma, Xi).}

\item{probaQuantile}{The probabilityof the value above which data are supposed to follow a GPD. A common value is 0.95.}

\item{nbDaysPerYear}{The number of days considered per year (integer).}

\item{nbYears}{The number of distinct years. Default value is 1.}

\item{h}{The parameter of non-concurrence (integer).}

\item{blockSizes}{Vector of integers of size 3. The sizes of the blocks considered for the declustering, with the univariates first and the bivariate block size at the end.}

\item{Dparam}{Cf documentation of the dgaps function of the exdex package. Default value is 3.}

\item{probaOccurrence}{The probability that the values are reached before the return period time. Default value to 0.63.}

\item{logic}{Either "AND" or "OR". Couples of bloxk maxima are retained if they both exceed their respective threshold in case of "AND", and if at least one of them exceed its respective threshold in case of "OR". Default value to "AND".}
}
\value{
A list with, in that order: the return level of first variable, the return level of the second variable, chi and chiBarre.
}
\description{
Select either the biGPD approach or the copula approach to get return levels from a given bivariate return period.
}
