% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Loire_API_ERA5_1992_2021}
\alias{Loire_API_ERA5_1992_2021}
\title{Loire API time series from ERA5 (1992-2021)}
\format{
A numeric array with 1 dimension and 1830 observations.
}
\source{
ERA5 reanalysis.
}
\usage{
Loire_API_ERA5_1992_2021
}
\description{
A one-dimensional numeric array containing API values for the Loire river region from the ERA5 reanalysis for the period 1992--2021.
}
\keyword{datasets}
