% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biGPD.R
\name{ProbaEGPD}
\alias{ProbaEGPD}
\title{Compute the probability to be under the return level value with an EGPD.}
\usage{
ProbaEGPD(returnLevel, EGPDtype, EGPDParams, probaZero)
}
\arguments{
\item{returnLevel}{The value at which the cdf is computed.}

\item{EGPDtype}{1 or 4. See Naveau et al. (2016) and the package mev for further information.}

\item{EGPDParams}{The EGPD parameters. For type 1, it is (Kappa, Sigma, Xi), for type 4, it is (Prob, Kappa, Delta, Sigma, Xi).}

\item{probaZero}{The probability to be exactly zero.}
}
\value{
The cdf value at the return level value for an EGPD.
}
\description{
Compute the probability to be under the return level value with an EGPD.
}
