% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biGPD.R
\name{ReturnPeriodBiGPDReturnLevels}
\alias{ReturnPeriodBiGPDReturnLevels}
\title{Bivariate return period when searching for return levels.}
\usage{
ReturnPeriodBiGPDReturnLevels(
  probaRL,
  probaQuantile,
  FbarU1U2,
  extremalIndex1,
  extremalIndex2,
  extremalIndexBivOG,
  h,
  nbDaysPerYear,
  probaOccurrence
)
}
\arguments{
\item{h}{The parameter of non-concurrence (integer).}

\item{nbDaysPerYear}{The number of days considered per year (integer).}

\item{probaOccurrence}{The probability that the values are reached before the return period time. Default value to 0.63.}

\item{probas}{A vector of size 3, with the 2 univariate probabilities first and then the probability of bivariate exceedence. The first two probabilities are likely close to 1, whereas the third one is likely close to 0.}

\item{extremalIndexes}{A vector of size 3, with the 2 univariate extremal indexes first and then the bivariate extremal index.}
}
\value{
The bivariate return period.
}
\description{
Bivariate return period when searching for return levels.
}
