% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Total_Precipitation_ERA5_1992_2021}
\alias{Total_Precipitation_ERA5_1992_2021}
\title{Total precipitation from ERA5 (1992-2021)}
\format{
A numeric array with 1 dimension and 2760 observations.
}
\source{
ERA5 reanalysis.
}
\usage{
Total_Precipitation_ERA5_1992_2021
}
\description{
A one-dimensional numeric array containing total precipitation values from the ERA5 reanalysis for the period 1992--2021.
}
\keyword{datasets}
