% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{UnivariateDeclustering}
\alias{UnivariateDeclustering}
\title{Decluster univariate data.}
\usage{
UnivariateDeclustering(data, nbDaysPerYear, nbYears, threshold, blockSize)
}
\arguments{
\item{data}{A vector of data.}

\item{nbDaysPerYear}{The number of days considered per year (integer).}

\item{nbYears}{The number of distinct years.}

\item{threshold}{The value above which data are supposed to follow a GPD.}

\item{blockSize}{The size of the blocks used for declustering.}
}
\value{
A list of declustered data: only the maxima of blocks that also exceeds the threshold.
}
\description{
Decluster univariate data.
}
