% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{UnivariateReturnPeriodCopula}
\alias{UnivariateReturnPeriodCopula}
\title{Calculate the univariate return period with the GPD approach}
\usage{
UnivariateReturnPeriodCopula(
  proba,
  extremalIndex,
  nbDaysPerYear,
  probaQuantile,
  h,
  probaOccurrence
)
}
\arguments{
\item{proba}{The GPD probability corresponding to the return level.}

\item{extremalIndex}{The extremal index cf UnivariateExtremalIndex.}

\item{nbDaysPerYear}{The number of days considered per year (integer).}

\item{probaQuantile}{The probabilityof the value above which data are supposed to follow a GPD. A common value is 0.95.}

\item{h}{The parameter of non-concurrence (integer).}

\item{probaOccurrence}{The probability that the values are reached before the return period time. Default value to 0.63.}
}
\value{
The univariate return period with the GPD approach.
}
\description{
Calculate the univariate return period with the GPD approach
}
