% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biGPD.R
\name{ValueEGPD}
\alias{ValueEGPD}
\title{Compute the value corresponding to the cdf probability with an EGPD.}
\usage{
ValueEGPD(proba, EGPDtype, EGPDParams, probaZero)
}
\arguments{
\item{proba}{The probability at which the inverse cdf is computed.}

\item{EGPDtype}{1 or 4. See Naveau et al. (2016) and the package mev for further information.}

\item{EGPDParams}{The EGPD parameters. For type 1, it is (Kappa, Sigma, Xi), for type 4, it is (Prob, Kappa, Delta, Sigma, Xi).}

\item{probaZero}{The probability to be exactly zero.}
}
\value{
The value at which the cdf equals proba for an EGPD.
}
\description{
Compute the value corresponding to the cdf probability with an EGPD.
}
