% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{UnivariateReturnPeriodCopula}
\alias{UnivariateReturnPeriodCopula}
\title{Calculate the univariate return period with the GPD approach}
\usage{
UnivariateReturnPeriodCopula(
  returnLevel,
  GPDParam,
  extremalIndex,
  nbDaysPerYear,
  probaQuantile,
  h,
  probaOccurrence
)
}
\arguments{
\item{returnLevel}{The value defining the return period.}

\item{GPDParam}{The parameters of the GPD. In a list with the following order: c(threshold, sigma, xi).}

\item{extremalIndex}{The extremal index cf UnivariateExtremalIndex.}

\item{nbDaysPerYear}{The number of days considered per year (integer).}

\item{probaQuantile}{The probabilityof the value above which data are supposed to follow a GPD. A common value is 0.95.}

\item{h}{The parameter of non-concurrence (integer).}
}
\value{
The univariate return period with the GPD approach.
}
\description{
Calculate the univariate return period with the GPD approach
}
